(Try to) Fix vulnerability
