--- src/test/java/com/kawa/api/package-info.java
@@ -1,11 +1,0 @@
-/**
- * <b>Root</b> <code>package</code>.
- * <br>
- *
- * The purpose of this package is to provide the classes used to execute the
- * unit tests of this project.
- * <hr>
- *
- * @since 0.1.0 hydrogen
- */
-package com.kawa.api;
