+++ src/test/java/com/kawa/api/suits/HelloTests.java
@@ -1,4 +1,4 @@
-package com.kawa.api;
+package com.kawa.api.suits;
 
 import java.net.URL;
