+++ src/test/java/com/kawa/api/suits/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * <b>Test Suit</b> <code>package</code>.
+ * <br>
+ *
+ * The purpose of this package is to provide the classes used to execute the
+ * unit test suits of this project.
+ * <hr>
+ *
+ * @since 0.1.0 hydrogen
+ */
+package com.kawa.api.suits;
